convert Button to TS
